requirements.txt
absl-py==2.3.1
amqp==5.3.1
asgiref==3.10.0
astunparse==1.6.3
attrs==25.4.0
billiard==4.2.3
brotli==1.2.0
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect2==0.8.0
Django==5.2.8
django-celery==3.1.17
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
flatbuffers==25.12.19
fonttools==4.61.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.13.0
kombu==5.5.4
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.4.0
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
protobuf==6.33.2
pycparser==2.23
pydyf==0.12.1
Pygments==2.19.2
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tinycss2==1.5.1
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
weasyprint==67.0
webencodings==0.5.1
Werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
zopfli==0.4.0
